+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fixed bugs:
   * "Run Therion" button blinking red on every start.
   * Status bar message not updating on option edit.
+  * Duplicated point appearing after scrap.
 * Infrastructure maintenance:
   * Created MPCommandAux.addTHIDToElement().
   * Upgrade actions/checkout github action from v4 to v6.
@@ -18,6 +19,7 @@
   * flutter pub upgrade --major-versions.
   * Making copyWith do a deep copy for options map, children, etc.
   * Testing point duplication.
+  * Default positionInParent on MPCommandFactory.addElements() changed to mpAddChildAtEndMinusOneOfParentChildrenList.
 
 ## 0.3.0 - 2026-03-09 - The [Great Fanfarra](https://youtu.be/HlV59UbrY-E) release
 * New features: